+++ security/rbac/RouteRbac.js
@@ -42,7 +42,7 @@ module.exports = class RouteRbac extends Base {
             return next(new Forbidden);
         }
         user.setReturnUrl(req.originalUrl);
-        const url = user.module.get('urlManager').resolve(user.getLoginUrl(), user.module.getPathName());
+        const url = user.module.get('urlManager').resolve(user.getLoginUrl(), user.module.getRouteName());
         return res.redirect(url);
     }